qq_talk_main/pyproject.toml
[project]
name = "qq_talk"
version = "0.1.0"
description = "qq_talk"
readme = "README.md"
requires-python = ">=3.9, <4.0"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
plugin_dirs = ["plugins","src.plugins.ai_chat"]
builtin_plugins = []

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.0.0"
nonebot-adapter-napcat = "^0.2.0"
httpx = "^0.23.0"
aiohttp = "^3.8.0"
